added solutions to comments. moved api logic to custom hook
