+++ src/components/CountriesList/CountriesList.tsx
@@ -1,22 +1,12 @@
-import { useState, useEffect } from "react";
+import { useEffect } from "react";
 import { StyledDisplaySection } from "./CountriesList.styles";
-import api from "../../api/countriesApi";
-import { CountriesProps } from "./CountriesProps";
-import { SingleCountry } from "../singleCountry/SingleCountry";
+import { CountryCard } from "../countryCard/CountryCard";
+import { useFetchCountries } from "../../api/useFetchCountries";
 
 export const CountriesList = () => {
-	const [error, setError] = useState<boolean>();
-	const [countries, setCountries] = useState<CountriesProps[]>();
+	const { error, countries, fetchCountries } = useFetchCountries();
 
 	useEffect(() => {
-		const fetchCountries = async () => {
-			try {
-				const response = await api.get("/all");
-				setCountries(response.data);
-			} catch (err: unknown) {
-				setError(true);
-			}
-		};
 		fetchCountries();
 	}, []);
 
@@ -24,17 +14,16 @@ export const CountriesList = () => {
 		<>
 			{error && <p>There is an error fetching API.</p>}
 			<StyledDisplaySection>
-				{countries
-					? countries.map(({ name, flags, population, region, capital }) => (
-							<SingleCountry
-								name={name}
-								flags={flags}
-								population={population}
-								region={region}
-								capital={capital}
-							/>
-					  ))
-					: "loading"}
+				{!countries && "loading"}
+				{countries?.map(({ name, flags, population, region, capital }) => (
+					<CountryCard
+						name={name}
+						flags={flags}
+						population={population}
+						region={region}
+						capital={capital}
+					/>
+				))}
 			</StyledDisplaySection>
 		</>
 	);
