+++ src/components/countryCard/CountryCard.styles.ts
@@ -1,10 +1,10 @@
 import styled from "styled-components";
 
-export const StyledDivWithCountryData = styled.div`
+export const StyledCountryCard = styled.div`
 	margin: 5px;
 	display: flex;
 	justify-content: space-between;
 	flex-direction: column;
-	max-width: 8rem;
-	max-height: 10rem;
+	max-width: 10rem;
+	max-height: 12rem;
 `;
