+++ src/components/countryCard/CountryCard.tsx
@@ -1,18 +1,18 @@
 import { CountriesProps } from "../CountriesList/CountriesProps";
-import { StyledDivWithCountryData } from "./SingleCountry.styles";
+import { StyledCountryCard } from "./CountryCard.styles";
 
-export const SingleCountry = ({
+export const CountryCard = ({
 	name,
 	flags,
 	population,
 	region,
 	capital,
 }: CountriesProps) => (
-	<StyledDivWithCountryData>
+	<StyledCountryCard>
 		<img src={flags.svg} alt={flags.alt} />
 		<p>{name.common}</p>
 		<p>Population:{population}</p>
 		<p>Region:{region}</p>
 		<p>Capital: {capital}</p>
-	</StyledDivWithCountryData>
+	</StyledCountryCard>
 );
